refactor : Member to MemberResponse
