+++ src/main/java/board/domain/member/Member.java
@@ -1,7 +1,6 @@
 package board.domain.member;
 
 
-import board.dto.member.MemberResponse;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -37,16 +36,4 @@ public class Member {
         this.age = age;
     }
 
-    public static MemberResponse toMemberResponse(Member member){
-        MemberResponse m = MemberResponse.builder()
-                .age(member.getAge())
-                .pw(member.getPw())
-                .interest(member.getInterest())
-                .id(member.getId())
-                .username(member.getUsername())
-                .address(member.getAddress())
-                .build();
-
-        return m;
-    }
 }
